Did population of the `stocks` and `price_history` tables by buying the Financial Modeling Prep (FMP) Starter plan api, enabling access to the `/v3/sp500_constituent` endpoint. Successfully populated the `stocks` table with 503 S&P 500 constituents and updated the `/stocks/price-history/populate-all` endpoint to fetch historical price data for each stock over the past year, including current `marketCap` and `peRatio` values via the `/quote` endpoint. This resulted in approximately 126,756 entries in the `price_history` table. Addressed challenges related to API rate limits, data parsing, and model alignment to ensure accurate data integration.

+++ src/main/java/com/PortfolioHeatmap/models/FMPQuoteResponse.java
@@ -36,7 +36,7 @@ public class FMPQuoteResponse {
     private Double pe;
 
     @JsonProperty("marketCap")
-    private Double marketCap;
+    private Long marketCap;
 
     // Getters and Setters
 
@@ -108,11 +108,11 @@ public class FMPQuoteResponse {
         this.pe = pe;
     }
 
-    public Double getMarketCap() {
+    public Long getMarketCap() {
         return marketCap;
     }
 
-    public void setMarketCap(Double marketCap) {
+    public void setMarketCap(Long marketCap) {
         this.marketCap = marketCap;
     }
